feat: define components for each page of the sidebar panel
