+++ packages/web/src/components/session-menu/secrets-panel.tsx
@@ -1,0 +1,13 @@
+// import { SessionMenuPanelContentsProps } from '.';
+
+type PropsType = {}; // Pick<SessionMenuPanelContentsProps, 'session' | 'openDepsInstallModal'>;
+
+export default function SessionMenuPanelSecrets(_props: PropsType) {
+  return (
+    <>
+      <div>Secrets</div>
+
+      TODO
+    </>
+  );
+}
